# app/models.py
# generated by datamodel-codegen:
#   filename:  openapi.yaml
#   timestamp: 2026-01-05T16:42:32+00:00

from __future__ import annotations

from typing import Any

from pydantic import BaseModel


class UploadResponse(BaseModel):
    task_id: str


class ImageStats(BaseModel):
    width: int | None = None
    height: int | None = None
    mean_color: list[float] | None = None
    hog_features: list[float] | None = None
    ai_probability: float | None = None
    hog_image_url: str | None = None
    fractal_dimension: float | None = None
    fractal_dim_small: float | None = None
    fractal_dim_large: float | None = None


class AggregateStats(BaseModel):
    total_images: int
    avg_width: float | None = None
    avg_height: float | None = None
    avg_color: list[float] | None = None


class Result(BaseModel):
    id: str | None = None
    url: str | None = None
    stats: ImageStats | None = None


class TaskStatus(BaseModel):
    status: str
    progress: int
    steps: list[str]
    current_step: str | None = None
    completed_steps: list[str]
    partial_results: dict[str, Any] | None = None
    result: Result | None = None
    error: str | None = None